backend/pyproject.toml
[project]
name = "agente-inteligente-backend"
version = "0.1.0"
description = "Backend con arquitectura hexagonal para sistema de agentes"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "pgvector>=0.2.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "groq>=0.5.0",
    "litellm>=1.35.0",
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    "fpdf2>=2.7.0",
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "python-multipart>=0.0.7",
    "passlib[bcrypt]>=1.7.4",
    "PyJWT>=2.8.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"